requirements.txt
click==8.1.7
colorama==0.4.6
et-xmlfile==1.1.0
h11==0.14.0
numpy==1.26.2
openpyxl==3.1.2
pandas==2.1.3
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
tzdata==2023.3
uvicorn==0.24.0.post1
